llm-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
huggingface-hub==0.24.6
llama-cpp-python==0.3.16
pydantic==2.8.2
sentence-transformers>=2.2.0
torch>=2.1.0
numpy>=1.24.0
